+++ R/udunits.R
@@ -1,31 +1,50 @@
-#' Test if two units are convertible
+#' \pkg{udunits2} utilities
 #'
-#' Parses and checks whether units can be converted by UDUNITS-2. Units may not
-#' be convertible either because they are different magnitudes or because one
-#' (or both) units are not defined in the database.
+#' Some \pkg{udunits2} utilities are exposed to the user. These functions are
+#' useful for checking whether units are convertible or converting between units
+#' without having to create \pkg{units} objects.
 #'
-#' @param x character or object of class \code{symbolic_units}, for the symbol
-#' of the first unit.
-#' @param y character or object of class \code{symbolic_units}, for the symbol
-#' of the second unit.
+#' @param from character or object of class \code{symbolic_units},
+#' for the symbol of the original unit.
+#' @param to   character or object of class \code{symbolic_units},
+#' for the symbol of the unit to convert.
 #'
-#' @return boolean, \code{TRUE} if both units exist and are convertible.
+#' @return \code{ud_are_convertible}
+#' returns \code{TRUE} if both units exist and are convertible,
+#' \code{FALSE} otherwise.
+#'
+#' @name udunits2
+#' @export
+ud_are_convertible <- function(from, to) {
+  ud_convertible(ud_char(from), ud_char(to))
+}
+
+#' @param x numeric vector
+#'
+#' @return \code{ud_convert}
+#' returns a numeric vector with \code{x} converted to new unit.
+#'
+#' @name udunits2
 #' @export
 #'
 #' @examples
 #' ud_are_convertible("m", "km")
+#' ud_convert(100, "m", "km")
+#'
 #' a <- set_units(1:3, m/s)
 #' ud_are_convertible(units(a), "km/h")
-#' ud_are_convertible("s", "kg")
-ud_are_convertible = function(x, y) {
-	stopifnot(inherits(x, c("character", "symbolic_units")), inherits(y, c("character", "symbolic_units")))
-	res <- try(R_ut_are_convertible(
-	  R_ut_parse(ud_char(x)), R_ut_parse(ud_char(y))), silent = TRUE)
-	! inherits(res, "try-error") && res
+#' ud_convert(1:3, units(a), "km/h")
+#'
+#' ud_are_convertible("degF", "degC")
+#' ud_convert(32, "degF", "degC")
+ud_convert <- function(x, from, to) {
+  ud_convert_doubles(x, ud_char(from), ud_char(to))
 }
 
 ud_char <- function(x) {
   if (is.character(x)) return(x)
+  if (!inherits(x, "symbolic_units")) stop("not a unit")
+
   res <- if (length(x$numerator))
     paste(x$numerator, collapse=" ") else "1"
   if (length(x$denominator))