+++ NEWS.md
@@ -27,6 +27,8 @@
 * Improve `keep_units()` helper for more general usage scenarios;
   #394 @d-morrison addressing #392
 
+* Add `ud_convert()` to convert units of a vector; #399 @dlebauer addressing #398
+
 # version 0.8-5
 
 * avoid -Wformat-security warning on CRAN